+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-# 0.0.1 - WIP
+# Changes
+
+## 0.1.0 - WIP
+
+- update TODO's
+- update main workflow
+- add basic router
+- add home and error views
+
+## 0.0.1
 
 - todo's